(MOT-4433) refactor(testkit): split provider contract modules
